Refactor to multiple spec files
